refactor: ActorSessionRandomStore 클래스 경로 변경
